+++ packages/icons/bin/kit.cjs
